pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "codexify"
version = "0.1.0"
description = "Python CLI & Module to intelligently compile project/code trees into a single text output, ideal for generating context for Large Language Models (LLMs)."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
  {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["llm", "codegen", "context", "compiler", "developer-tools", "python", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Compilers",
    "Topic :: Text Processing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "PyYAML>=5.0",
    "tiktoken>=0.3.0",
    "gitignore-parser>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/codexify"
Repository = "https://github.com/yourusername/codexify"

[project.scripts]
codexify = "codexify.cli:run_cli"

[tool.setuptools] # Required for include/exclude with setuptools
packages = ["codexify", "codexify.core"]
# Or use find directive if you prefer:
# [tool.setuptools.packages.find]
# where = ["."]
# include = ["codexify*"]
# exclude = ["tests*", "examples*"]